requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
cffi==1.11.5
colorama==0.3.9
cryptography==2.3.1
dj-database-url==0.5.0
Django==2.0.6
django-appconf==1.0.2
django-filter==1.1.0
django-imagekit==4.0.2
django-ranged-response==0.2.0
django-registration-redux==2.4
django-reviews==1.2.1
django-simple-captcha==0.5.9
django-storages==1.6.6
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
flake8==3.5.0
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
pilkit==2.0
Pillow==5.0.0
psycopg2==2.7.5
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.6.4
pylint==1.9.2
pytz==2018.4
six==1.11.0
virtualenv==16.0.0
whitenoise==3.3.1
wrapt==1.10.11
